fix(wizard): Refactored to accommodate a CloudForms use case
After speaking with J. Phillips, Cloud Forms has a use case where the user has clicked next, but is blocked from navigating to the next step/substep until a process has completed. I’ve modified the example and wizard component as follows.

The step/substep allowNavAway config property is initially set to false, which keeps the user on the current step. When the user clicks the wizard’s next button, an event is emitted which an app developer can use to kick off a process. The app developer would then show a progress spinner in the step/substep main panel while the user is waiting. When the process has completed, allowNavAway is set to true and the wizard’s goToNextStep() method is called, allowing navigation to the next step/substep.