fix(release): bumped patternfly-eng-release for latest build changes
In order to support semantic-release, 'npm publish' must be run from the root instead of the $DIST_DIR
directory. When publishing a sub folder, npm loses the ability to insert the correct gitHead information, which
prevents semantic-release from working properly.

This solution copies the contents of the dist directory to the root directory. The .npmignore file lists
anything in the root that should not be published.