chore(release): clean up npm packaging
This greatly simplifies how we npm package the dist directory. Until now, the entire contents of dist directory had been copied to the root directory.

This was done in order to support semantic-release, because ’npm publish' must be run from the root instead of the dist directory. When publishing a sub folder, npm loses the ability to insert the correct gitHead information, which prevents semantic-release from working properly.

The dist path is provided via the module and typings properties of package.json. And the files property of package.json defines the directory to be published. This allows patternfly-ng modules to be imported as normal, without needed to specify a path.