fix(sparkline): avoid regenerating chart on every data update
When only chart data has changed and config remains unchanged, the sparkline chart component will
not be regenerated but the data simply reloaded instead. This should be a lighter-weight operation
and avoids a visual flicker and redraw of the chart component on data update.

fixes https://github.com/patternfly/patternfly-ng/issues/275