feat(vertical-nav): add IDs to primary, secondary, and tertiary links
Following the patternfly-ng convention for adding IDs with components like list, copy, etc. We're generating unique IDs based on the ID given to the component. Since the vertical nav is a list of items, the default primary, secondary, and tertiary links (output for item 0) will look like this:

<a id="pfng-vertical-navigation1-primary-nav-item0">
<a id="pfng-vertical-navigation1-secondary-nav-item0">
<a id="pfng-vertical-navigation1-tertiary-nav-item0">

Note that developers have the option to override the default ID (e.g., pfng-vertical-navigation) we use as a prefix. For example, if you give the component an "myNav" ID, those links would look more like this.

<a id="myNav-primary-nav-item0">
<a id="myNav-secondary-nav-item0">
<a id="myNav-tertiary-nav-item0">